training/requirements.txt
numpy
torch
tqdm
pytorch-crf
khmersegment
regex